+++ src/NewRelic.OpenTelemetry/NewRelicExporterOptions.cs
@@ -13,12 +13,16 @@ namespace NewRelic.OpenTelemetry
     /// </summary>
     public class NewRelicExporterOptions
     {
+        internal NewRelicExporterOptions()
+        {
+        }
+
         /// <summary>
         /// REQUIRED: Your Insights Insert API Key.  This value is required in order to communicate with the
         /// New Relic Endpoint. 
         /// </summary>
         /// <see cref="https://docs.newrelic.com/docs/insights/insights-data-sources/custom-data/introduction-event-api#register">for more information.</see>
-        public string? ApiKey
+        public string ApiKey
         {
             get => TelemetryConfiguration.ApiKey;
             set => TelemetryConfiguration.ApiKey = value;
@@ -27,29 +31,10 @@ namespace NewRelic.OpenTelemetry
         /// <summary>
         /// The New Relic endpoint where information is sent.
         /// </summary>
-        public Uri EndpointUrl
+        public Uri Endpoint
         {
             get => TelemetryConfiguration.TraceUrl;
             set => TelemetryConfiguration.TraceUrl = value;
-        }
-
-        /// <summary>
-        /// Logs messages sent-to and received-by the New Relic endpoints.  This setting
-        /// is useful for troubleshooting, but is not recommended in production environments.
-        /// </summary>
-        public bool AuditLoggingEnabled
-        {
-            get => TelemetryConfiguration.AuditLoggingEnabled;
-            set => TelemetryConfiguration.AuditLoggingEnabled = value;
-        }
-
-        /// <summary>
-        /// Identifies the name of a service for which information is being reported to New Relic.
-        /// </summary>
-        public string? ServiceName
-        {
-            get => TelemetryConfiguration.ServiceName;
-            set => TelemetryConfiguration.ServiceName = value;
         }
 
         /// <summary>
@@ -62,6 +47,9 @@ namespace NewRelic.OpenTelemetry
         /// </summary>
         public BatchExportProcessorOptions<Activity> BatchExportProcessorOptions { get; set; } = new BatchExportProcessorOptions<Activity>();
 
-        internal TelemetryConfiguration TelemetryConfiguration { get; } = new TelemetryConfiguration();
+        internal TelemetryConfiguration TelemetryConfiguration { get; } = new TelemetryConfiguration()
+        {
+            ServiceName = "OpenTelemetry Exporter",
+        };
     }
 }