+++ tests/NewRelic.Telemetry.Tests/ConfigTests.cs
@@ -26,7 +26,7 @@ namespace NewRelic.Telemetry.Tests
         private const string ApiKeyProdValue = "987654";
         private const string ApiKeyNewRelicValue = "123456";
         private const string ApiKeyDiffProductvalue = "ABCDEFG";
-        private const string ApiKeyDefaultValue = null;
+        private const string ApiKeyDefaultValue = "";
 
         private const int SendTimeoutSecondsDiffProdValue = 500;
 
@@ -35,7 +35,7 @@ namespace NewRelic.Telemetry.Tests
         private const bool AuditLoggingEnabledDiffProdValue = true;
 
         private const string ServiceNameNewRelicValue = NRConfigSection + "Service";
-        private const string ServiceNameDefaultValue = null;
+        private const string ServiceNameDefaultValue = "New Relic Telemetry SDK";
 
         private const int BackoffMaxSecondsDefaultValue = 80;